src/test/java/ru/didyk/coinkeeper/controller/AccountControllerTest.java
package ru.didyk.coinkeeper.controller;

import com.fasterxml.jackson.databind.ObjectMapper;
import org.junit.jupiter.api.Test;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
import org.springframework.boot.test.context.SpringBootTest;
import org.springframework.http.MediaType;
import org.springframework.test.context.jdbc.Sql;
import org.springframework.test.web.servlet.MockMvc;
import ru.didyk.coinkeeper.model.UserCategory;

import static org.assertj.core.api.Assertions.assertThat;
import static org.hamcrest.Matchers.is;
import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.*;
import static org.springframework.test.web.servlet.result.MockMvcResultHandlers.print;
import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.jsonPath;
import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;

@SpringBootTest
@AutoConfigureMockMvc
@Sql(value = {"/account-before.sql"}, executionPhase = Sql.ExecutionPhase.BEFORE_TEST_METHOD)
class AccountControllerTest {

    @Autowired
    private AccountController accountController;

    @Autowired
    private MockMvc mockMvc;

    private final ObjectMapper mapper = new ObjectMapper();

    @Test
    public void contextLoads() {
        assertThat(accountController).isNotNull();
    }

    @Test
    void getAccount() throws Exception {
        mockMvc.perform(get("/api/coinkeeper/{id}", 1L))
                .andDo(print())
                .andExpect(status().isOk())
                .andExpect(jsonPath("$.id", is(1)))
                .andExpect(jsonPath("$.balance", is(10000)));
    }

    @Test
    void addAccount() throws Exception {
        UserCategory userCategory = new UserCategory();
        userCategory.setId(1L);

        String requestJson = mapper.writeValueAsString(userCategory);

        mockMvc.perform(post(AccountController.PATH + "account")
                        .content(requestJson)
                        .contentType(MediaType.APPLICATION_JSON))
                .andExpect(status().isOk())
                .andExpect(jsonPath("$.id", is(1)));
    }

    @Test
    void updateAccount() throws Exception {
        UserCategory userCategory = new UserCategory();
        userCategory.setId(1L);

        String requestJson = mapper.writeValueAsString(userCategory);

        mockMvc.perform(post(AccountController.PATH + "account")
                        .content(requestJson)
                        .contentType(MediaType.APPLICATION_JSON))
                .andExpect(status().isOk())
                .andExpect(jsonPath("$.id", is(1)));
    }

    @Test
    void deleteAccount() throws Exception {

        mockMvc.perform(delete(AccountController.PATH + "delete/" + "{id}", 1)
                        .contentType(MediaType.APPLICATION_JSON))
                .andExpect(status().isOk());
    }
}